Test home stage

+++ src/main/java/by/itacademy/demyankov/Calculator.java
@@ -2,7 +2,7 @@ package by.itacademy.demyankov;
 
 public class Calculator {
 
-    public int add(int a, int b) {
+    public double add(double a, double b) {
         return a + b;
     }
